requirements.txt
python-magic
yara-python==4.3.1 ; platform_system != "Windows"

